added delete button

+++ frontend/src/components/posts/PostDropDown.jsx
@@ -1,22 +1,30 @@
 import { useState } from "react";
-import { useDispatch } from "react-redux";
+import { useDispatch, useSelector } from "react-redux";
 import { deletePost } from "../../store/reducers/post";
 import "./PostDropDown.css";
 import PostModal from "./PostModal";
-const PostDropDown = ({ postId }) => {
+import * as modalActions from "../../store/reducers/modals";
+
+const PostDropDown = ({ postId, postUserId }) => {
   const dispatch = useDispatch();
+  const sessionUser = useSelector((state) => state.session.user);
   const [showDropDown, setShowDropDown] = useState(false);
   const handleDropDownToggle = () => [setShowDropDown(!showDropDown)];
+  const modalType = useSelector((state) => state.modals.type === "SHOW_MODAL");
 
   const handleDelete = () => {
-    dispatch(deletePost(postId));
-    setShowDropDown(false);
+    if (sessionUser && sessionUser.id === postUserId) {
+      dispatch(deletePost(postId));
+      setShowDropDown(false);
+    }
   };
 
   const handleEdit = () => {
     // dispatch(updatePost(postId));
-    <PostModal />;
-    setShowDropDown(false);
+    if (sessionUser.id === postUserId) {
+      dispatch(modalActions.showModal("SHOW_MODAL", { postId }));
+      setShowDropDown(false);
+    }
   };
 
   return (
@@ -26,8 +34,16 @@ const PostDropDown = ({ postId }) => {
       </button>
       {showDropDown && (
         <div className="dropDownOptions">
-          <button onClick={handleEdit}>Edit Post</button>
-          <button onClick={handleDelete}>Delete Post</button>
+          {sessionUser && sessionUser.id === postUserId && (
+            <>
+              <button onClick={handleEdit}>Edit Post</button>
+              {modalType && (
+                <PostModal userName={sessionUser.userName} postId={postId} />
+              )}
+
+              <button onClick={handleDelete}>Delete Post</button>
+            </>
+          )}
         </div>
       )}
     </div>
